Fix for false notes being played on some MIDI devices.
Some MIDI device's do not have a Note Off event built into it, but instead once a note is released it sends a second Note On event with a velocity of 0. I hadn't accounted for this, and as a result some users were getting doubled keypresses. This fixes the issue.